+++ rust-sdk/whirlpool/src/account.rs
@@ -46,10 +46,15 @@ pub(crate) async fn get_token_accounts_for_owner(
     let mut token_accounts: Vec<ParsedTokenAccount> = Vec::new();
     for account in accounts {
         if let UiAccountData::Json(data) = account.account.data {
+            let token_program = match data.program.as_str() {
+                "spl-token" => &spl_token::ID.to_string(),
+                "spl-token-2022" => &spl_token_2022::ID.to_string(),
+                pubkey => pubkey,
+            };
             let token: Parsed = from_value(data.parsed)?;
             token_accounts.push(ParsedTokenAccount {
                 pubkey: Pubkey::from_str(&account.pubkey)?,
-                token_program: Pubkey::from_str(&data.program)?,
+                token_program: Pubkey::from_str(token_program)?,
                 mint: Pubkey::from_str(&token.info.mint)?,
                 amount: token.info.token_amount.amount.parse::<u64>()?,
             });