+++ ts-sdk/tx-sender/src/computeBudget.ts
@@ -2,27 +2,29 @@ import {
   getSetComputeUnitPriceInstruction,
   getSetComputeUnitLimitInstruction,
 } from "@solana-program/compute-budget";
-import {
-  prependTransactionMessageInstruction,
+import type {
   IInstruction,
-  isWritableRole,
   MicroLamports,
   Address,
   Slot,
 } from "@solana/web3.js";
+import {
+  prependTransactionMessageInstruction,
+  isWritableRole,
+} from "@solana/web3.js";
 import { rpcFromUrl } from "./compatibility";
+import type { Percentile } from "./config";
 import {
   DEFAULT_COMPUTE_UNIT_MARGIN_MULTIPLIER,
-  Percentile,
   getPriorityFeeConfig,
   getComputeUnitMarginMultiplier,
   getRpcConfig,
 } from "./config";
-import { TxMessage } from "./priorityFees";
+import type { TxMessage } from "./priorityFees";
 
 export async function processComputeBudgetForTxMessage(
   message: TxMessage,
-  computeUnits: number
+  computeUnits: number,
 ) {
   const { rpcUrl, supportsPriorityFeePercentile } = getRpcConfig();
   const priorityFee = getPriorityFeeConfig();
@@ -36,7 +38,7 @@ export async function processComputeBudgetForTxMessage(
       message.instructions,
       rpcUrl,
       supportsPriorityFeePercentile,
-      priorityFee.priorityFeePercentile ?? "50"
+      priorityFee.priorityFeePercentile ?? "50",
     );
 
     if (!priorityFee.maxCapLamports) {
@@ -57,7 +59,7 @@ export async function processComputeBudgetForTxMessage(
       getSetComputeUnitPriceInstruction({
         microLamports: priorityFeeMicroLamports,
       }),
-      message
+      message,
     );
   }
   message = prependTransactionMessageInstruction(
@@ -68,7 +70,7 @@ export async function processComputeBudgetForTxMessage(
           DEFAULT_COMPUTE_UNIT_MARGIN_MULTIPLIER)
       ),
     }),
-    message
+    message,
   );
 
   return message;
@@ -90,14 +92,14 @@ async function calculateDynamicPriorityFees(
   instructions: readonly IInstruction[],
   rpcUrl: string,
   supportsPercentile: boolean,
-  percentile: Percentile
+  percentile: Percentile,
 ) {
   const writableAccounts = getWritableAccounts(instructions);
   if (supportsPercentile) {
     return await getRecentPrioritizationFeesWithPercentile(
       rpcUrl,
       writableAccounts,
-      percentile
+      percentile,
     );
   } else {
     const rpc = rpcFromUrl(rpcUrl);
@@ -108,18 +110,6 @@ async function calculateDynamicPriorityFees(
       .filter((pf) => pf.prioritizationFee > 0)
       .map((pf) => pf.prioritizationFee);
     const sorted = nonZero.sort((a, b) => Number(a - b));
-
-    if (percentile === "50") {
-      const mid = sorted.length / 2;
-      if (sorted.length === 0) return BigInt(0);
-      if (sorted.length % 2 === 0) {
-        return (
-          (sorted[Math.floor(mid - 1)] + sorted[Math.floor(mid)]) / BigInt(2)
-        );
-      } else {
-        return sorted[Math.floor(mid)];
-      }
-    }
     return (
       sorted[
       Math.floor(sorted.length * (percentileNumber(percentile) / 100))
@@ -131,7 +121,7 @@ async function calculateDynamicPriorityFees(
 async function getRecentPrioritizationFeesWithPercentile(
   rpcEndpoint: string,
   writableAccounts: Address[],
-  percentile: Percentile
+  percentile: Percentile,
 ) {
   const response = await fetch(rpcEndpoint, {
     method: "POST",