+++ rust-sdk/integration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orca-so/whirlpools-rust-integration
 
+## 0.2.0
+
+### Minor Changes
+
+- [#726](https://github.com/orca-so/whirlpools/pull/726) [`7f0ca73`](https://github.com/orca-so/whirlpools/commit/7f0ca73f49ce8354bb9156bba326cd5d9e93d665) Thanks [@wjthieme](https://github.com/wjthieme)! - Add support for solana v2 crates
+
 ## 0.1.1
 
 ### Patch Changes